Factor out base keystone types
