+++ .changeset/lucky-masks-enjoy.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/types': patch
+---
+
+Refactored types to isolate base keystone type definitions.
